\name{cheeses}
\alias{cheeses}
\docType{data}
\title{
cheeses
}

\usage{data(cheeses)}
\format{
  A data frame with 288 observations on the following 16 variables.
  \describe{
    \item{\code{ProductCode}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F}}
    \item{\code{SubjectCode}}{a factor with levels \code{S5733} \code{S5734} \code{S5735} \code{S5736} \code{S5737} \code{S5738} \code{S5739} \code{S5740} \code{S5741} \code{S5742} \code{S5743} \code{S5744} \code{S5745} \code{S5746} \code{S5747} \code{S5748}}
    \item{\code{Replicate}}{a factor with levels \code{R1} \code{R2} \code{R3}}
    \item{\code{Session}}{a factor with levels \code{Unknown}}
    \item{\code{Butter}}{a numeric vector}
    \item{\code{Crumbly}}{a numeric vector}
    \item{\code{Dense}}{a numeric vector}
    \item{\code{Elastic}}{a numeric vector}
    \item{\code{Fat}}{a numeric vector}
    \item{\code{Fruity}}{a numeric vector}
    \item{\code{Hard}}{a numeric vector}
    \item{\code{Intense}}{a numeric vector}
    \item{\code{Milky}}{a numeric vector}
    \item{\code{Salted}}{a numeric vector}
    \item{\code{Smooth}}{a numeric vector}
    \item{\code{Sticky}}{a numeric vector}
  }
}
\examples{
data(cheeses)
## maybe str(cheeses) ; plot(cheeses) ...
}
\keyword{datasets}
