% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL_PlotPCACVA.R
\name{INTERNAL_PlotPCACVA}
\alias{INTERNAL_PlotPCACVA}
\title{INTERNAL_PlotPCACVA}
\usage{
INTERNAL_PlotPCACVA(
  variables,
  individuals,
  suppIndividualsToPlot = FALSE,
  suppIndividuals = NULL,
  biplot = TRUE,
  variablesColors = NULL,
  individualsColors = NULL,
  suppIndividualsColors = NULL,
  variablesLabels = TRUE,
  individualsLabels = TRUE,
  suppIndividualsLabels = FALSE,
  individualsEllipses = NULL,
  individualsSegments = NULL,
  xlab = "",
  ylab = "",
  expand = NULL,
  mainTitle = NULL,
  subTitle = NULL,
  returnX = FALSE,
  returnY = FALSE,
  listEllipsesByRep = NULL,
  cex = 0.8,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{variables}{matrix of variables}

\item{individuals}{matrix of individuals}

\item{suppIndividualsToPlot}{TRUE or FALSE}

\item{suppIndividuals}{matrix of suppIndividuals to use}

\item{biplot}{TRUE by default, but can be FALSE}

\item{variablesColors}{vector of colors for variables}

\item{individualsColors}{vector of colors for variables}

\item{suppIndividualsColors}{vector of colors for supplementary individuals}

\item{variablesLabels}{TRUE  if to be displayed, FALSE otherwise}

\item{individualsLabels}{TRUE  if to be displayed, FALSE otherwise}

\item{suppIndividualsLabels}{TRUE  if to be displayed, FALSE otherwise}

\item{individualsEllipses}{array containing ellipses coordinates}

\item{individualsSegments}{TRUE or FALSE}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{expand}{useful for biplot. Number indicating how to expand the graph}

\item{mainTitle}{title of the graph}

\item{subTitle}{subtitle of the graph}

\item{returnX}{FALSE by default. Boolean indicating if the x axis should be reverted}

\item{returnY}{FALSE by default. Boolean indicating if the y axis should be reverted}

\item{listEllipsesByRep}{internal parameters NULL by default}

\item{cex}{size of text}

\item{xlim}{vector of two numbers indicating the limits of x axis}

\item{ylim}{vector of two numbers indicating the limits of y axis
confInt   limit for the confidence ellipses. By default 0.9}
}
\description{
INTERNAL_PlotPCACVA
}
