% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCA.R
\name{PlotPCA}
\alias{PlotPCA}
\title{PlotPCA}
\usage{
PlotPCA(
  res.PCA,
  panellists = "None",
  representation = "DistanceBiplot",
  nbAxes = "Auto",
  title = "PCA",
  confInt = 0.95,
  ellipsesType = "Barycentric",
  productColors = NULL,
  returnX = FALSE,
  returnY = FALSE,
  type = "R",
  fileName = "PCA",
  ellipsesCalculation = "Chi",
  expand = NULL,
  variablesLabels = TRUE,
  variablesColors = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{res.PCA}{results of the function PCA}

\item{panellists}{representation of the panellists on the map: "None" if the panellist are not represented on the map, "labels" if they are represented by their names, or "points" if they are represented by points}

\item{representation}{"DistanceBiplot","TwoMaps","CorrelationBiplot".}

\item{nbAxes}{"Auto"}

\item{title}{custumizes the title of your PCA}

\item{confInt}{the confidence limit for the ellipses}

\item{ellipsesType}{"barycentric" for the confidence ellipses around the product means "individual" for the representation of a proportion of confInt of the panelist scores}

\item{productColors}{vector of colors for products}

\item{returnX}{reverses the abscisse axis}

\item{returnY}{reverses the ordinate axis}

\item{type}{"R" for a R graph, "png","pdf","jpeg" for registering the graph in these formats in your working directory}

\item{fileName}{name of your file if you choose type="png","pdf" or "jpeg"}

\item{ellipsesCalculation}{"Chi", "F" or "Sas"}

\item{expand}{for biplot}

\item{variablesLabels}{TRUE by default.}

\item{variablesColors}{vector of colors for variables}

\item{xlim}{vector of two numbers indicating the limit of x axis}

\item{ylim}{vector of two numbers indicating the limit of y axis}
}
\description{
PlotPCA
}
\examples{
{
 data(cheeses)
 PlotPCA(PCA(cheeses,productName="ProductCode",subjectName="SubjectCode",
 replicate="Replicate",sessionName="Session"))
}
}
